+++ cli/src/pcluster/schemas/cluster_schema.py
@@ -35,6 +35,7 @@ from pcluster.config.cluster_config import (
     AwsBatchQueueNetworking,
     AwsBatchScheduling,
     AwsBatchSettings,
+    CapacityReservationTarget,
     CapacityType,
     CloudWatchDashboards,
     CloudWatchLogs,
@@ -946,6 +947,32 @@ class TimeoutsSchema(BaseSchema):
         return Timeouts(**data)
 
 
+class CapacityReservationTargetSchema(BaseSchema):
+    """Represent the schema of the CapacityReservationTarget section."""
+
+    capacity_reservation_id = fields.Str(metadata={"update_policy": UpdatePolicy.QUEUE_UPDATE_STRATEGY})
+    capacity_reservation_resource_group_arn = fields.Str(metadata={"update_policy": UpdatePolicy.QUEUE_UPDATE_STRATEGY})
+
+    @post_load()
+    def make_resource(self, data, **kwargs):
+        """Generate resource."""
+        return CapacityReservationTarget(**data)
+
+    @validates_schema
+    def no_coexist_instance_type_flexibility(self, data, **kwargs):
+        """Validate that 'capacity_reservation_id' and 'capacity_reservation_resource_group_arn' do not co-exist."""
+        if self.fields_coexist(
+            data,
+            ["capacity_reservation_id", "capacity_reservation_resource_group_arn"],
+            one_required=True,
+            **kwargs,
+        ):
+            raise ValidationError(
+                "A Capacity Reservation Target needs to specify either Capacity Reservation ID or "
+                "Capacity Reservation Resource Group ARN."
+            )
+
+
 class ClusterDevSettingsSchema(BaseDevSettingsSchema):
     """Represent the schema of Dev Setting."""
 
@@ -1116,6 +1143,9 @@ class SlurmComputeResourceSchema(_ComputeResourceSchema):
     efa = fields.Nested(EfaSchema, metadata={"update_policy": UpdatePolicy.QUEUE_UPDATE_STRATEGY})
     disable_simultaneous_multithreading = fields.Bool(metadata={"update_policy": UpdatePolicy.COMPUTE_FLEET_STOP})
     schedulable_memory = fields.Int(metadata={"update_policy": UpdatePolicy.QUEUE_UPDATE_STRATEGY})
+    capacity_reservation_target = fields.Nested(
+        CapacityReservationTargetSchema, metadata={"update_policy": UpdatePolicy.QUEUE_UPDATE_STRATEGY}
+    )
 
     @validates_schema
     def no_coexist_instance_type_flexibility(self, data, **kwargs):
@@ -1206,6 +1236,9 @@ class _CommonQueueSchema(BaseQueueSchema):
     )
     iam = fields.Nested(QueueIamSchema, metadata={"update_policy": UpdatePolicy.SUPPORTED})
     image = fields.Nested(QueueImageSchema, metadata={"update_policy": UpdatePolicy.QUEUE_UPDATE_STRATEGY})
+    capacity_reservation_target = fields.Nested(
+        CapacityReservationTargetSchema, metadata={"update_policy": UpdatePolicy.QUEUE_UPDATE_STRATEGY}
+    )
 
 
 class SlurmQueueSchema(_CommonQueueSchema):